feat(svelte): add rune-based test infrastructure with real   $state/$effect
  Replace mock-based Svelte tests with .svelte.test.ts files that run through
  the Svelte compiler, enabling real $state and $effect verification. This
  catches rune wiring bugs (incorrect $effect dependencies, $state proxy
  behaviour) that the mock-only approach missed.

  - Add vitest.config.svelte.ts (jsdom env, svelte plugin)
  - Add jsdom dev dependency
  - Migrate context.test.ts → context.svelte.test.ts (7 tests)
  - Migrate use-all.test.ts → use-all.svelte.test.ts (9 tests)
  - Exclude .svelte.test.ts from the main vitest config
