+++ packages/jazz-tools/vitest.config.svelte.ts
@@ -1,0 +1,10 @@
+import { defineConfig } from "vitest/config";
+import { svelte } from "@sveltejs/vite-plugin-svelte";
+
+export default defineConfig({
+  plugins: [svelte({ hot: false })],
+  test: {
+    environment: "jsdom",
+    include: ["src/**/*.svelte.test.ts"],
+  },
+});
